// src/services/tts/types.ts
export type TTSGranularity = 'sentence' | 'word';

export type TTSMediaMetadataMode = 'sentence' | 'paragraph' | 'chapter';

export type TTSHighlightOptions = {
  style: 'highlight' | 'underline' | 'squiggly' | 'outline';
  color: string;
};

export type TTSVoice = {
  id: string;
  name: string;
  lang: string;
  disabled?: boolean;
};

export type TTSVoicesGroup = {
  id: string;
  name: string;
  voices: TTSVoice[];
  disabled?: boolean;
};

export type TTSMark = {
  offset: number;
  name: string;
  text: string;
  language: string;
};